Protect apis with nextauth

+++ pages/api/v1/trade/delete-trade-by-id/index.ts
@@ -7,15 +7,16 @@ import dbConnect from 'utils/dbConnect'
 
 import Trade from 'models/Trade'
 
-export default async function (req: NextApiRequest, res: NextApiResponse): Promise<void> {
+export default async function (req: NextApiRequest, res: NextApiResponse): Promise<unknown> {
   try {
-    await dbConnect()
     const session = await getSession({ req })
+    if (!session) return res.status(401)
+
+    await dbConnect()
     const { tradeId } = req.body
     const deletedTrade = await Trade.findOneAndDelete({ _id: tradeId, user: session.user._id })
     res.status(200).json(deletedTrade)
   } catch (error) {
-    if (error) throw error
-    res.status(400).json('error')
+    res.status(400).json(error)
   }
 }
